refactor: extract indexer.rs tests to indexer_tests.rs
Moves the 1608-line mod tests block into a separate file to keep
indexer.rs focused on production code only (514 lines).
